Use transaction-emitter-lib from transaction-emitter
