+++ crates/transaction-emitter-lib/src/args.rs
@@ -1,19 +1,19 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
+use std::convert::TryFrom;
+
+use anyhow::{bail, format_err, Result};
 use aptos_config::keys::ConfigKey;
 use aptos_crypto::ed25519::Ed25519PrivateKey;
 use aptos_sdk::types::chain_id::ChainId;
 use clap::Parser;
 use url::Url;
 
-#[derive(Parser, Debug)]
-pub struct ClusterArgs {
-    /// Nodes the cluster should connect to, e.g. http://node.mysite.com:8080
-    /// If the port is not provided, it is assumed to be 8080.
-    #[clap(short, long, required = true, min_values = 1)]
-    pub targets: Vec<Url>,
+const DEFAULT_API_PORT: u16 = 8080;
 
+#[derive(Parser, Debug)]
+pub struct MintArgs {
     /// Ed25519PrivateKey for minting coins
     #[clap(long, parse(try_from_str = ConfigKey::from_encoded_string))]
     pub mint_key: Option<ConfigKey<Ed25519PrivateKey>>,
@@ -23,10 +23,21 @@ pub struct ClusterArgs {
 
     #[clap(long, default_value = "TESTING")]
     pub chain_id: ChainId,
+}
+
+#[derive(Parser, Debug)]
+pub struct ClusterArgs {
+    /// Nodes the cluster should connect to, e.g. http://node.mysite.com:8080
+    /// If the port is not provided, it is assumed to be 8080.
+    #[clap(short, long, required = true, min_values = 1, parse(try_from_str = parse_target))]
+    pub targets: Vec<Url>,
 
     /// If set, try to use public peers instead of localhost.
     #[clap(long)]
     pub vasp: bool,
+
+    #[clap(flatten)]
+    pub mint_args: MintArgs,
 }
 
 #[derive(Parser, Debug)]
@@ -44,16 +55,10 @@ pub struct EmitArgs {
     pub burst: bool,
 
     /// This can only be set in conjunction with --burst. By default, when burst
-    /// is enabled, we do not ever check the transaction stats. If this is set,
-    /// we will check the stats once at the end.
-    #[clap(long, requires = "burst")]
-    pub check_stats_at_end: bool,
-
-    /// The transaction emitter will submit no more than this many transactions
-    /// per second. So if max TPS is 1600 and there are 16 workers, each worker
-    /// will submit no more than 100 per second each.
-    #[clap(long, default_value = "100000")]
-    pub max_tps: u64,
+    /// is enabled, we check stats once at the end of the emitter run. If you
+    /// would like to opt out of that, you can use this flag.
+    #[structopt(long, requires = "burst")]
+    pub do_not_check_stats_at_end: bool,
 
     #[clap(long, default_value = "30")]
     pub txn_expiration_time_secs: u64,
@@ -64,4 +69,24 @@ pub struct EmitArgs {
 
     #[clap(long, help = "Percentage of invalid txs", default_value = "0")]
     pub invalid_tx: usize,
+}
+
+fn parse_target(target: &str) -> Result<Url> {
+    let mut url = Url::try_from(target).map_err(|e| {
+        format_err!(
+            "Failed to parse listen address, try adding a scheme, e.g. http://: {}",
+            e
+        )
+    })?;
+    if url.scheme().is_empty() {
+        bail!("Scheme must not be empty, try prefixing URL with http://");
+    }
+    if url.port().is_none() {
+        url.set_port(Some(DEFAULT_API_PORT)).map_err(|_| {
+            anyhow::anyhow!(
+                "Failed to set port to default value, make sure you have set a scheme like http://"
+            )
+        })?;
+    }
+    Ok(url)
 }
