+++ crates/transaction-emitter-lib/src/wrappers.rs
@@ -1,29 +1,24 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::args::{ClusterArgs, EmitArgs};
-use crate::cluster::Cluster;
-use crate::emit::EmitJobRequest;
-use crate::emit::EmitThreadParams;
-use crate::emit::TxnEmitter;
-use crate::emit::TxnStats;
-use crate::instance::Instance;
+use crate::{
+    args::{ClusterArgs, EmitArgs},
+    cluster::Cluster,
+    emit::{EmitJobRequest, EmitThreadParams, TxnEmitter, TxnStats},
+    instance::Instance,
+};
 use anyhow::{Context, Result};
 use aptos_sdk::transaction_builder::TransactionFactory;
-use rand::rngs::StdRng;
-use rand::Rng;
-use rand_core::OsRng;
-use rand_core::SeedableRng;
-use std::cmp::min;
-use std::convert::TryFrom;
-use std::time::Duration;
+use rand::{rngs::StdRng, Rng};
+use rand_core::{OsRng, SeedableRng};
+use std::{cmp::min, convert::TryFrom, time::Duration};
 
 pub async fn emit_transactions(
     cluster_args: &ClusterArgs,
     emit_args: &EmitArgs,
 ) -> Result<TxnStats> {
     let cluster = Cluster::try_from(cluster_args).context("Failed to build cluster")?;
-    emit_transactions_with_cluster(&cluster, &emit_args, cluster_args.vasp).await
+    emit_transactions_with_cluster(&cluster, emit_args, cluster_args.vasp).await
 }
 
 pub async fn emit_transactions_with_cluster(
@@ -35,7 +30,7 @@ pub async fn emit_transactions_with_cluster(
         wait_millis: args.wait_millis,
         wait_committed: !args.burst,
         txn_expiration_time_secs: args.txn_expiration_time_secs,
-        check_stats_at_end: args.check_stats_at_end,
+        do_not_check_stats_at_end: args.do_not_check_stats_at_end,
     };
     let duration = Duration::from_secs(args.duration);
     let client = cluster.random_instance().rest_client();
@@ -53,7 +48,6 @@ pub async fn emit_transactions_with_cluster(
             .accounts_per_client(args.accounts_per_client)
             .thread_params(thread_params)
             .invalid_transaction_ratio(args.invalid_tx)
-            .max_tps(args.max_tps)
             .gas_price(1);
     if let Some(workers_per_endpoint) = args.workers_per_ac {
         emit_job_request = emit_job_request.workers_per_endpoint(workers_per_endpoint);
