+++ x.toml
@@ -77,7 +77,6 @@ workspace-members = [
 
     # Don't include these since they contain test features that we don't want in production.
     "forge",
-    "transaction-emitter",
 ]
 
 [hakari.final-excludes]
@@ -194,7 +193,6 @@ members = [
     "peer-monitoring-service-server", # This will be removed once the peer monitoring service is plugged in
     "peer-monitoring-service-types", # This will be removed once the peer monitoring service is plugged in
     "smoke-test",
-    "transaction-emitter",
     "x",
     "x-core",
     "x-lint",
